add ban module
